fix state persistance and refactor init store
