--- packages/psss/src/createReducers.js
@@ -1,10 +1,0 @@
-/*
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-import { auth, weeklyReports } from './store';
-
-export const createReducers = () => ({
-  auth,
-  weeklyReports,
-});
